requirements.txt
requests
python-whois
dnspython
reportlab
Pillow
beautifulsoup4
ttkthemes